+++ packages/synthetix-governance/contracts/submodules/election/DebtShareManager.sol
@@ -13,9 +13,9 @@ contract DebtShareManager is ElectionBase, DebtShareStorage {
     error DebtShareSnapshotIdNotSet();
 
     event DebtShareContractSet(address contractAddress);
-    event DebtShareSnapshotIdSet(uint128 snapshotId);
+    event DebtShareSnapshotIdSet(uint snapshotId);
 
-    function _setDebtShareSnapshotId(uint128 snapshotId) internal {
+    function _setDebtShareSnapshotId(uint snapshotId) internal {
         DebtShareStore storage store = _debtShareStore();
 
         uint currentEpochIndex = _getCurrentEpochIndex();
@@ -24,10 +24,10 @@ contract DebtShareManager is ElectionBase, DebtShareStorage {
         emit DebtShareSnapshotIdSet(snapshotId);
     }
 
-    function _getDebtShareSnapshotId() internal view returns (uint128) {
+    function _getDebtShareSnapshotId() internal view returns (uint) {
         DebtShareStore storage store = _debtShareStore();
 
-        uint128 debtShareId = store.debtShareIds[_getCurrentEpochIndex()];
+        uint debtShareId = store.debtShareIds[_getCurrentEpochIndex()];
         if (debtShareId == 0) {
             revert DebtShareSnapshotIdNotSet();
         }
@@ -56,7 +56,7 @@ contract DebtShareManager is ElectionBase, DebtShareStorage {
     function _getDebtShare(address user) internal view returns (uint) {
         DebtShareStore storage store = _debtShareStore();
 
-        uint128 debtShareId = store.debtShareIds[_getCurrentEpochIndex()];
+        uint debtShareId = store.debtShareIds[_getCurrentEpochIndex()];
 
         return store.debtShareContract.balanceOfOnPeriod(user, debtShareId);
     }
